requirements.txt
absl-py==2.1.0
aenum==3.1.15
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
assemblyai==0.27.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.6.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
config==0.5.1
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.7
decorator==4.4.2
deepgram-sdk==3.3.1
distlib==0.3.7
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
docstring_parser==0.16
eventlet==0.36.1
ffmpeg-python==0.2.0
filelock==3.12.4
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.0
future==1.0.0
google-ai-generativelanguage==0.6.6
google-api-core==2.15.0
google-api-python-client==2.134.0
google-auth==2.26.1
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.59.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-language==2.13.3
google-cloud-resource-manager==1.12.4
google-cloud-storage==2.18.0
google-cloud-vision==3.5.0
google-crc32c==1.5.0
google-generativeai==0.7.2
google-resumable-media==2.7.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
idna==3.4
imageio==2.34.2
imageio-ffmpeg==0.5.1
itsdangerous==2.1.2
jax==0.4.29
jaxlib==0.4.29
Jinja2==3.1.2
kiwisolver==1.4.5
lxml==5.2.2
magic-filter==1.0.12
MarkupSafe==2.1.3
marshmallow==3.21.3
matplotlib==3.9.0
mediapipe==0.10.14
ml-dtypes==0.4.0
moviepy==1.0.3
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.4
openai==0.28.1
opencv-contrib-python==4.10.0.82
opencv-python==4.10.0.82
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.2
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.10.0
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.2
pyTelegramBotAPI==4.15.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
python-telegram-bot==20.7
pytz==2024.1
requests==2.31.0
rsa==4.9
scipy==1.13.1
shapely==2.0.5
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.7
soupsieve==2.5
SQLAlchemy==2.0.29
telebot==0.0.5
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.0.5
virtualenv==20.24.5
websockets==12.0
Werkzeug==2.3.7
wsproto==1.2.0
yarg==0.1.9
yarl==1.9.2
